dynamic style with edit water rates and computation

+++ database/migrations/2019_09_20_015458_create_rates_table.php
@@ -15,8 +15,8 @@ class CreateRatesTable extends Migration
     {
         Schema::create('rates', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->double('from');
-            $table->double('to');
+            $table->integer('from');
+            $table->integer('to');
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
